+++ HISTORY.rst
@@ -2,10 +2,12 @@
 History
 =======
 
-0.4.2 (2023-10-20)
+UNRELEASED
 ------------------
-* Add base model for `ImportJob` and `ExportJob`
+* Add base model for ``ImportJob`` and ``ExportJob``
 * Extend import results template: show validation errors in table
+* Add force-import feature: skip rows with errors while importing
+* Add ``skip_parse_step`` parameter for importing API
 
 0.4.1 (2023-09-25)
 ------------------